==> 직장인을 위한 데이터분석 실무 with 파이썬/03 Tourist Event (1) data preprocessing.ipynb <==
### 데이터 전처리

import pandas as pd

# header = 1 : 두번째 줄에 칼럼명이 있음
# usecols = 'A:G' : A~G칼럼만 사용
# skipfooter = 4 : 아래에서 4줄은 사용하지 않음

kto_201901 = pd.read_excel('data/3_Tourists_Event/files/kto_201901.xlsx',\
                           header=1, usecols = 'A:G', skipfooter = 4)
kto_201901.head()

kto_201901.tail()

#### 데이터 탐색

kto_201901.info()

kto_201901.describe()

# 데이터 값이 0인 레이블 확인
conditions = (kto_201901['관광'] == 0) | \
             (kto_201901['상용'] == 0) | \
             (kto_201901['공용'] == 0) | \
             (kto_201901['유학/연수'] == 0)

kto_201901[conditions]

#### 데이터 프레임에 기준연월 추가

kto_201901['기준연월'] = '2019-01'
kto_201901.head()

#### 국적 데이터만 남기기

# 칼럼이 어떤 정보로 이루어졌는지 확인하는 메쏘드 unique()
kto_201901['국적'].unique()

# 대륙 데이터를 리스트로 만듬
continents_list = ['아시아주', '미주', '구주', '대양주', '아프리카주', '기타대륙', '교포소계']
continents_list

# continents_list가 아닌 데이터만 선택
kto_201901_country = kto_201901[kto_201901['국적'].isin(continents_list)==False]
kto_201901_country['국적'].unique()

kto_201901_country.head()

# 인덱스 값 초기화
kto_201901_country_newindex = kto_201901_country.reset_index(drop = True)
kto_201901_country_newindex.head()

#### 대륙 칼럼 만들기

continents = ['아시아']*25 + ['아메리카']*5 + ['유럽']*23 + ['오세아니아']*3 + ['아프리카']*2 + ['기타대륙'] + ['교포']
print(continents)

kto_201901_country_newindex['대륙'] = continents
kto_201901_country_newindex.head()

kto_201901_country_newindex.tail()

#### 국적별 관광객 비율 살펴보기

kto_201901_country_newindex['관광객비율(%)'] = \
  round(kto_201901_country_newindex['관광'] / kto_201901_country_newindex['계'] * 100,1)
kto_201901_country_newindex.head()

# 관광객비율 칼럼을 내림차순으로 정렬하여 상위 5개국 확인
kto_201901_country_newindex.sort_values(by = '관광객비율(%)', ascending=False).head()

# 관광객비율 칼럼을 오름차순으로 정렬하여 하위 5개국 확인
kto_201901_country_newindex.sort_values(by = '관광객비율(%)', ascending=True).head()

kto_201901_country_newindex.pivot_table(values = '관광객비율(%)', index='대륙', aggfunc='mean')

# 중국의 비율 확인
kto_201901_country_newindex[kto_201901_country_newindex['국적'] == '중국']

#### 기준년월별로 전체 외국인 관광객 대비 국적별 관광객 비율 살펴보기

tourist_sum = sum(kto_201901_country_newindex['관광'])
tourist_sum

kto_201901_country_newindex['전체비율(%)'] = \
   round(kto_201901_country_newindex['관광'] / tourist_sum * 100, 1)
kto_201901_country_newindex.sort_values(by='전체비율(%)', ascending = False).head()

### 데이터 전처리과정 함수로 만들기

def create_kto_data(yy,mm):
    # 1. 불러올 엑셀 파일 경로를 지정
    file_path = f'data/3_Tourists_Event/files/kto_{yy}{mm}.xlsx'
    
    # 2. 엑셀 파일 불러오기
    df = pd.read_excel(file_path, header=1, skipfooter=4, usecols='A:G')
    
    # 3. '기준년월' 칼럼 추가
    df['기준년월'] = f'{yy}-{mm}'
    
    # 4. '국적'칼럼에서 대륙 제거하고 국가만 남기기
    ignore_list = ['아시아주', '미주', '구주', '대양주', '아프리카주','기타대륙', '교포소계']
    df_country = df[df['국적'].isin(ignore_list)  == False].reset_index(drop=True)
    
    # 5. '대륙'칼럼 추가
    continents = ['아시아']*25 + ['아메리카']*5 + ['유럽']*23 + ['오세아니아']*3 + ['아프리카']*2 + ['기타대륙'] + ['교포']
    df_country['대륙'] = continents
    
    # 6. 국가별 '관광객비율(%)' 칼럼 추가
    df_country['관광객비율(%)'] = round(df_country['관광'] / df_country['계'] *100 ,1)
    
    # 7. '전체비율(%)' 칼럼 추가
    tourist_sum = sum(df_country['관광'])
    df_country['전체비율(%)'] = round(df_country['관광'] / tourist_sum * 100, 1)
    
    # 8. 결과출력
    return(df_country)

kto_test = create_kto_data(2018,12)
kto_test.head()

### 반복문을 통해 다수의 엑셀 데이터를 불러와서 합치기

for yy in range(2010,2020):
    for mm in range(1,13):
        # mm을 1이 아닌 01로 만들기 위해 zfill 메쏘드 활용
        # zfill 메쏘드는 문자열에만 적용가능
        mm_str = str(mm).zfill(2)
        yymm = f'{yy}{mm_str}'
        print(yymm)

# 데이터를 담을 빈 데이터프레임 만들기
df = pd.DataFrame()

for yy in range(2010,2020):
    for mm in range(1,13):
        temp = create_kto_data(str(yy), str(mm).zfill(2))
        df = df.append(temp, ignore_index = True)

# 데이터 중 201909 엑셀이 없어서 생기는 오류
df.head()

df.tail()

for yy in range(2010,2020):
    for mm in range(1,13):
        # try 이하 구문을 실행
        try:
            temp = create_kto_data(str(yy), str(mm).zfill(2))
            df = df.append(temp, ignore_index = True)
        # 만약 try 구문에서 오류 발생시 except 이하 구문 실행
        except:
            pass

df.tail()

df.info()

#### 통합테이터를 엑셀파일로 저장

df.to_excel('data/3_Tourists_Event/files/kto_total.xlsx',index=False)
