==> 직장인을 위한 데이터분석 실무 with 파이썬/04 Hotplace (1) crawling.ipynb <==
### 크롤링

#### 인스타그램의 검색 결과 URL을 만드는 함수

def insta_searching(word):
    url = "https://www.instagram.com/explore/tags/" + word
    return url

#### 인스타그램에 로그인하는 함수

def insta_login(email,password):
    input_id = driver.find_elements_by_css_selector('input._2hvTZ.pexuQ.zyHYP')[0]
    input_id.clear()
    input_id.send_keys(email)
    
    input_pw = driver.find_elements_by_css_selector('input._2hvTZ.pexuQ.zyHYP')[1]
    input_pw.clear()
    input_pw.send_keys(password)
    input_pw.submit()
    time.sleep(3)
    
    save_info = driver.find_element_by_css_selector('#react-root > section > main > div > div > div > section > div > button')
    save_info.click()
    time.sleep(3)
    
    alarm = driver.find_element_by_css_selector('body > div.RnEpo.Yx5HN > div > div > div > div.mt3GC > button.aOOlW.HoLwm')
    alarm.click()

#### 첫번째 게시글 열기

def select_first(driver):
    
    # find_element_by_css_selector : BeutifulSoup을 사용하지 않고 Selenium에서 직접 html에 접근하는 메쏘드
    first = driver.find_element_by_css_selector('div._9AhH0')
    
    # 게시물 클릭
    first.click()
    
    # 정보 로딩에 시간이 필요하므로 3초간 대기
    time.sleep(3)

#### 게시글 정보 가져오기

# 정규 표현식을 위한 클래스
import re

def get_content(driver):
    # 1. 현재 페이지의 HTML정보 가져오기
    html = driver.page_source
    soup = BeautifulSoup(html, 'lxml')
    
    # 2. 본문 내용 가져오기
    try:
        content = soup.select('div.C4VMK > span')[0].text
    except:
        content = ''
        
    # 3. 본문 내용에서 해시태그 가져오기(정규표현식 활용)
    tags = re.findall(r'#[^\s#,\\]+', content)
    
    # 4. 작성일자 정보 가져오기
    date = soup.select('time.FH9sR.Nzb55')[0]['datetime'][:10]
    
    # 5. 좋아요 수 가져오기
    try:
        like = soup.select('div.Nm9Fw > button')[0].text[4:-1]
    except:
        like = 0
        
    # 6. 위치정보 가져오기
    try:
        place = soup.select('div.JF9hh > a.O4GlU')[0].text
    except:
        place = ''
    
    # 7. 수집한 정보 저장하기
    data = [content, date, like, place, tags]
    return data

# 다음 게시글을 여는 함수

def move_next(driver):
    right = driver.find_element_by_css_selector('a._65Bje.coreSpriteRightPaginationArrow')
    right.click()
    time.sleep(3)

### 크롤링 과정 통합 함수

from selenium import webdriver
from bs4 import BeautifulSoup
import time
import re

# 1. 크롬브라우저 열기
driver = webdriver.Chrome('C:/Users/NK/Desktop/programming/python/chromedriver.exe')

# 2. 인스타그램에 접속
driver.get('https://www.instagram.com')
time.sleep(2)

# 3. 인스타그램 로그인
email = '01038067982'
pw = '!@nam9595'
insta_login(email,pw)

# 4. 인스타그램 검색페이지 URL 만들기
word = '제주여행' # 검색어
url = insta_searching(word)

# 5. 검색페이지 접속하기
driver.get(url)
time.sleep(3)

# 6. 첫 번째 게시글 열기
select_first(driver)

# 7. 비어있는 변수(results) 만들기
results = []

# 8. 여러 게시물 수집하기
target = 100
for i in range(target):
    data = get_content(driver)
    results.append(data)
    move_next(driver)
    
print(results[:2])

### 수집 데이터 저장

import pandas as pd

results_df = pd.DataFrame(results)
results_df.columns = ['content', 'data', 'like', 'place', 'tags']
results_df.to_excel('crawling_jejutravel.xlsx')

### 여러 엑셀파일의 중복을 제거한 후 통합 저장

# 여러 엑셀파일을 통합
jeju_insta_df = pd.DataFrame([])

folder = 'crawling_data/'
f_list = ['crawling_jeju.xlsx', 'crawling_jejumat.xlsx', 'crawling_jejutravel.xlsx']

for fname in f_list:
    fpath = folder + fname
    temp = pd.read_excel(fpath, usecols='B:F')
    jeju_insta_df = jeju_insta_df.append(temp)

jeju_insta_df.reset_index(drop=True, inplace=True)
jeju_insta_df.head()

# 중복데이터 삭제

jeju_insta_df.drop_duplicates(subset = ['content'], inplace = True)
print(jeju_insta_df.shape)

jeju_insta_df.to_excel('crawling_data/crawling_raw.xlsx', index = False)
