==> 직장인을 위한 데이터분석 실무 with 파이썬/05 Starbucks (1) data collecting.ipynb <==
### 스타벅스 입점데이터 크롤링

from selenium import webdriver
from bs4 import BeautifulSoup
import pandas as pd
import time

driver = webdriver.Chrome('C:/Users/NK/Desktop/programming/python/chromedriver.exe')
url = 'https://www.starbucks.co.kr/store/store_map.do?disp=locale'
driver.get(url)
time.sleep(3)

# '서울' > '전체' 버튼 클릭

seoul_btn = '#container > div > form > fieldset > div > section > article.find_store_cont > article > article:nth-child(4) > div.loca_step1 > div.loca_step1_cont > ul > li:nth-child(1) > a'
driver.find_element_by_css_selector(seoul_btn).click()
time.sleep(3)

all_btn = '#mCSB_2_container > ul > li:nth-child(1) > a'
driver.find_element_by_css_selector(all_btn).click()
time.sleep(3)

# HTML 파서 만들기

html = driver.page_source
soup = BeautifulSoup(html, 'html.parser')

# html 태그를 이용하여 각 매장 정보 가져오기

starbucks_soup_list = soup.select('li.quickResultLstCon')
print(len(starbucks_soup_list))

starbucks_soup_list[0]

# 매장 정보 샘플 확인
starbucks_store = starbucks_soup_list[0]
name = starbucks_store.select('strong')[0].text.strip()
lat = starbucks_store['data-lat'].strip()
lng = starbucks_store['data-long'].strip()
store_type = starbucks_store.select('i')[0]['class'][0][4:]
address = str(starbucks_store.select('p')[0]).split('<br/>')[0].split('>')[1]
tel = str(starbucks_store.select('p')[0]).split('<br/>')[1].split('<')[0]

print(name)
print(lat)
print(lng)
print(store_type)
print(address)
print(tel)

# 서울시 스타벅스 매장 목록 데이터 만들기

starbucks_list = []
for item in starbucks_soup_list:
    name = item.select('strong')[0].text.strip()
    lat = item['data-lat'].strip()
    lng = item['data-long'].strip()
    store_type = item.select('i')[0]['class'][0][4:]
    address = str(item.select('p')[0]).split('<br/>')[0].split('>')[1]
    tel = str(item.select('p')[0]).split('<br/>')[1].split('<')[0]
    
    starbucks_list.append([name, lat, lng, store_type, address, tel])

# starbucks_list를 데이터프레임으로 생성
columns = ['매장명', '위도', '경도', '매장타입', '주소', '전화번호']
seoul_starbucks_df = pd.DataFrame(starbucks_list, columns = columns)
seoul_starbucks_df.head()

seoul_starbucks_df.info()

# 엑셀로 저장
seoul_starbucks_df.to_excel('crawling_data/4_1_seoul_starbucks_list.xlsx', index=False)

### 공공데이터 수집

import requests
import pandas as pd

# 서울 열린데이터광장 API 호출 키
SEOUL_API_AUTH_KEY = '****************************'

# 시군구 목록을 가져오는 API 호출 URL
url = f'http://openapi.seoul.go.kr:8088/{SEOUL_API_AUTH_KEY}/json/SdeTlSccoSigW/1/25/'
print(url)

# 서울 열린데이터광장 OPEN API 호출
result_dict = requests.get(url).json()
result_dict

result_data = result_dict['SdeTlSccoSigW']

print(result_data['list_total_count'])
print(result_data['RESULT'])
print(result_data['row'][0])

data_list = result_data['row']
sample_df = pd.DataFrame(data_list)
sample_df.head()

#### 서울열린데이터광장 OPEN API를 호출하는 함수 작성

    서울 열린데이터광장 OPEN API는 동일한 구조이므로 함수로 작성하여 활용

# 서울열린데이터광장 API 호출 키

SEOUL_API_AUTH_KEY = '*******************'

# 서울열린데이터광장 OPEN API 호출을 위한 공통 함수
def seoul_open_api_data(url, service):
    data_list = None
    try:
        result_dict = requests.get(url).json()
        result_data = result_dict[service]
        code = result_data['RESULT']['CODE']
        if code == 'INFO-000':
            data_list = result_data['row']
    except:
        pass
    return data_list

#### 시군구 목록 데이터 수집후 엑셀 다운로드

# url주소 내 항목들은 서울열린데이터광장 open api탭에서 확인가능
sgg_url = f'http://openapi.seoul.go.kr:8088/{SEOUL_API_AUTH_KEY}/json/SdeTlSccoSigW/1/25/'

sgg_data_list = seoul_open_api_data(url, 'SdeTlSccoSigW')
sgg_data_list[0]

# 결과데이터를 pandas의 데이터프레임으로 변환
columns = ['SIG_CD', 'SIG_KOR_NM', 'LAT', 'LNG']
sgg_df = pd.DataFrame(data = sgg_data_list, columns = columns)
sgg_df.head()

sgg_df.columns = ['시군구코드', '시군구명', '위도', '경도']
sgg_df.head()

sgg_df.info()

# 데이터 엑셀로 저장

sgg_df.to_excel('crawling_data/4_2_seoul_sgg_list.xlsx', index = False)

### 공통함수를 이용 서울시 시군구별 인구데이터 가져오기

pop_url = f'http://openapi.seoul.go.kr:8088/{SEOUL_API_AUTH_KEY}/json/octastatapi419/1/26/'
pop_data_list = seoul_open_api_data(pop_url, 'octastatapi419')

sgg_pop_df = pd.DataFrame(data = pop_data_list)

sgg_pop_df.head()

sgg_pop_df.info()

# 필요없는 합계 데이터 제거
sgg_pop_df_selected = sgg_pop_df[sgg_pop_df['JACHIGU'] != '합계']
sgg_pop_df_selected.head()

# 분석에 필요한 칼럼만 선택
columns = ['JACHIGU', 'GYE_1']
sgg_pop_df_final = sgg_pop_df_selected[columns]
sgg_pop_df_final.columns = ['시군구명', '주민등록인구']
sgg_pop_df_final.head()

# 데이터를 엑셀파일로 저장
sgg_pop_df_final.to_excel('crawling_data/4_3_sgg_pop.xlsx',index = False)

#### 공통함수를 이용해 서울시 시군구별 사업체 데이터 가져오기

biz_url = f'http://openapi.seoul.go.kr:8088/{SEOUL_API_AUTH_KEY}/json/octastatapi104/1/450/'
biz_data_list = seoul_open_api_data(biz_url, 'octastatapi104')

sgg_biz_df = pd.DataFrame(biz_data_list)
sgg_biz_df.head(10)

sgg_biz_df_selected = sgg_biz_df[sgg_biz_df['DONG']=='소계']
sgg_biz_df_selected.head(10)

# 필요없는 칼럼 삭제 및 인덱스 초기화

columns = ['JACHIGU', 'GYE', 'SAEOPCHESU_1']
sgg_biz_df_final = sgg_biz_df_selected[columns]
sgg_biz_df_final.columns = ['시군구명', '종사자수', '사업체수']
sgg_biz_df_final = sgg_biz_df_final.reset_index(drop = True)

sgg_biz_df_final.head()

# 엑셀로 저장

sgg_biz_df_final.to_excel('crawling_data/4_3_sgg_biz.xlsx', index = False)
