==> 직장인을 위한 데이터분석 실무 with 파이썬/05 Starbucks (2) Data Preprocessing.ipynb <==
#### 데이터 로드

import pandas as pd

seoul_starbucks = pd.read_excel('crawling_data/4_1_seoul_starbucks_list.xlsx', header = 0)
seoul_starbucks.head()

# 스타벅스 주소에서 시군구명 정보 추출
sgg_names = []
for address in seoul_starbucks['주소']:
    sgg = address.split(' ')[1]
    sgg_names.append(sgg)

seoul_starbucks['시군구명'] = sgg_names
seoul_starbucks = seoul_starbucks[seoul_starbucks['시군구명'] != '67']
seoul_starbucks.head()

# 엑셀로 저장
seoul_starbucks.to_excel('crawling_data/4_4_seoul_starbucks_list.xlsx', index = False)

# 시군구 목록 데이터 불러오기
seoul_sgg = pd.read_excel('crawling_data/4_2_seoul_sgg_list.xlsx')
seoul_sgg.head()

# 시군구별 스타벅스 매장 수 세기
starbucks_sgg_count = seoul_starbucks.pivot_table(index = '시군구명', values = '매장명', aggfunc = 'count').rename(columns = {'매장명' : '스타벅스_매장수'})
starbucks_sgg_count.head()

seoul_sgg = pd.merge(seoul_sgg, starbucks_sgg_count, how = 'left', on = '시군구명')
seoul_sgg.head()

# 인구통계 데이터 불러오기
seoul_sgg_pop = pd.read_excel('crawling_data/4_3_sgg_pop.xlsx')
seoul_sgg_pop.head()

# 스타벅스 데이터와 인구통게 데이터 병합
seoul_sgg = pd.merge(seoul_sgg, seoul_sgg_pop, how='left', on='시군구명')
seoul_sgg.head()

# 서울시 시군구별 사업체 수 통계 데이터도 병합
seoul_sgg_biz = pd.read_excel('crawling_data/4_3_sgg_biz.xlsx')
seoul_sgg = pd.merge(seoul_sgg, seoul_sgg_biz, how = 'left', on = '시군구명')
seoul_sgg.head()

# 전처리 된 결과를 엑셀 파일로 저장
seoul_sgg.to_excel('crawling_data/4_5_seoul_sgg_stat.xlsx', index = False)
