==> 직장인을 위한 데이터분석 실무 with 파이썬/06 Best Product (1) Data Crawling.ipynb <==
### 데이터 수집 (1) 한 페이지 수집

#### 다나와 홈페이지 접속

from selenium import webdriver
from bs4 import BeautifulSoup
import time

driver = webdriver.Chrome('C:/Users/NK/Desktop/programming/python/chromedriver.exe')
url = 'http://search.danawa.com/dsearch.php?k1=무선청소기&module=goods&act=dispMain#'
driver.get(url)
time.sleep(3)

#### 상품 정보 가져오기

html = driver.page_source
soup = BeautifulSoup(html, 'html.parser')

# 1페이지에 대한 무선청소기 정보 가져오기

prod_items = soup.select('li.prod_item')
len(prod_items)

# 1페이지에 대한 무선청소기 정보 가져오기

prod_items = soup.select('ul.product_list > li.prod_item')
len(prod_items)

# 1페이지에 대한 무선청소기 정보 가져오기

prod_items = soup.select('div.main_prodlist > ul.product_list > li.prod_item')
len(prod_items)

# 첫번째 상품의 정보 가져오기

## 상품명
title = prod_items[0].select('a.click_log_product_standard_title_')[0].text.strip()

## 스펙
spec = prod_items[0].select('div.spec_list')[0].text.strip()

## 가격
price = prod_items[0].select('li.rank_one > p.price_sect > a > strong')[0].text.strip().replace(',','')

print(title)
print(spec)
print(price)

# 첫페이지의 모든 상품의 정보를 가져오는 반복문
import pandas as pd

prod_data = []

for item in prod_items:
    try:
        title = item.select('a.click_log_product_standard_title_')[0].text.strip()
    except:
        title = ''
    
    try:
        spec = item.select('div.spec_list')[0].text.strip()
    except:
        spec = ''
    
    try:
        price = item.select('li.rank_one > p.price_sect > a > strong')[0].text.strip().replace(',','')
    except:
        price = ''
        
    prod_data.append([title, spec, price])
    
print(len(prod_data))
pd.DataFrame(prod_data).head()

# 한 페이지의 검색결과를 수집하는 함수

def get_prod_items(prod_items):
    prod_data = []
    
    prod_data = []

    for item in prod_items:
        try:
            title = item.select('a.click_log_product_standard_title_')[0].text.strip()
        except:
            title = ''
    
        try:
            spec = item.select('div.spec_list')[0].text.strip()
        except:
            spec = ''
    
        try:
            price = item.select('li.rank_one > p.price_sect > a > strong')[0].text.strip().replace(',','')
            price = int(price)
        except:
            price = ''

        prod_data.append([title, spec, price])
        
    return prod_data

# 함수 테스트

prod_items = soup.select('div.main_prodlist > ul.product_list > li.prod_item')
prod_data = get_prod_items(prod_items)
print(len(prod_data))

### 데이터 수집 (2) 여러 페이지에 걸친 다나와 검색 페이지

# 검색페이지의 URL을 보면 query / page 만 바뀌는 것을 알 수 있음

# 다나와 검색 url을 만들어주는 함수
def get_search_page_url(keyword, page):
    return f'http://search.danawa.com/dsearch.php?query={keyword}&originalQuery={keyword}&volumeType=allvs&page={page}&limit=40&sort=saveDESC&list=list&boost=true&addDelivery=N&tab=goods'

keyword = '무선청소기'
page = '3'
url = get_search_page_url(keyword, page)
print(url)

### 주피터 노트북의 진행표시줄 처리

import time
from tqdm import tqdm_notebook

total_page = 10
for page in tqdm_notebook(range(1, total_page+1)):
    # 페이지 로딩 완료되기 위한 시간을 5초로 한다
    time.sleep(5)

### 여러 페이지에 걸친 상품 정보 수집

from selenium import webdriver
from bs4 import BeautifulSoup
from tqdm import tqdm_notebook

import pandas as pd
import time

driver = webdriver.Chrome('C:/Users/NK/Desktop/programming/python/chromedriver.exe')
driver.implicitly_wait(3)

keyword = '무선청소기'
total_page = 10
prod_data_total = []
# 진행 정도를 표현하는 tqdm을 적용
for page in tqdm_notebook(range(1, total_page + 1)):

    # 1. 검색페이지 이동
    url = get_search_page_url(keyword, page)
    driver.get(url)
    ## 페이지 로딩 완료까지 5초를 기다림
    time.sleep(5)
    
    # 2. 현재 페이지의 HTML 정보 가져오기
    html = driver.page_source
    soup = BeautifulSoup(html, 'html.parser')
    
    # 3. 상품 정보 추출
    prod_items = soup.select('div.main_prodlist > ul.product_list > li.prod_item')
    prod_item_list = get_prod_items(prod_items)
    
    # 4. 추출 데이터 저장
    prod_data_total += prod_item_list
    
len(prod_data_total)

# 수집 데이터 확인
data = pd.DataFrame(prod_data_total)
data.columns = ['상품명', '스펙 목록', '가격']
print(data.shape)
data.head()

# 수집 데이터 저장
data.to_excel('crawling_data/6_1_danawa_crawling_result.xlsx', index = False)
