==> 직장인을 위한 데이터분석 실무 with 파이썬/06 Best Product (2) Data Preprocessing.ipynb <==
import pandas as pd

data = pd.read_excel('crawling_data/6_1_danawa_crawling_result.xlsx')
data.info()
data.head()

#### 회사명, 모델명 정리

data['상품명'][:10]

company_list = []
product_list = []

for title in data['상품명']:
    title_info = title.split(' ',1)
    company_list.append(title_info[0])
    product_list.append(title_info[1])
    
print(len(company_list))
print(company_list[:5])
print(len(product_list))
print(product_list[:5])

#### 스펙 목록에서 카테고리, 사용시간, 흡입력 추출해서 정리

print(data['스펙 목록'][0],'\n')
print(data['스펙 목록'][1])

## 제품의 카테고리는 맨 앞에, 사용시간과 흡입력은 명시되어 있음

spec_list = data['스펙 목록'][0].split(' / ')
use_time_spec = []

for spec in spec_list:
    if '사용시간' in spec:
        use_time_spec = spec
    
print(use_time_spec)

# 스펙목록에서 원하는 원소 추출
category_list = []
use_time_list = []
suction_list = []

for spec_list in data['스펙 목록']:
    
    ## 1. 스펙 목록을 / 단위로 쪼갬
    spec_list = spec_list.split(' / ')

    ## 2. 스펙 목록의 첫번째 원소를 카테고리로 지정
    category_list.append(spec_list[0])
    
    ## 3. 사용시간 / 흡입력 정보가 없는 데이터를 위한 변수 생성
    use_time_value = None
    suction_value = None
    
    for spec in spec_list:
        ## 4. 스펙 목록에서 '사용시간'이라는 문자열이 포함된 원소를 사용시간으로 지정
        if '사용시간' in spec:
            use_time_value = spec.split(' ')[1]
        ## 5. 스펙 목록에서 '흡입력'이라는 문자열이 포함된 원소를 흡입력으로 지정 
        elif '흡입력' in spec:
            suction_value = spec.split(' ')[1]
    
    use_time_list.append(use_time_value)
    suction_list.append(suction_value)       

print('카테고리', len(category_list), category_list[:5])            
print('사용시간', len(use_time_list),use_time_list[:5])
print('흡입력', len(suction_list),suction_list[:5])

#### 사용시간 단위 통일시키기

use_time_list

# 사용시간을 분 단위로 조정하는 함수
def convert_time_minute(time):
    try:
        if '시간' in time:
            hour = time.split('시간')[0]
            if '분' in time:
                minute = time.split('시간')[-1].split('분')[0]
            else:
                minute = 0
        else:
            hour = 0
            minute = time.split('분')[0]
        
        return int(hour)*60 + int(minute)

    except:
        return None

# 함수 테스트
times = ['40분', '4분', '1시간', '3시간30분', '4시간']
for time in times:
    time_value = convert_time_minute(time)
    print(time, '=', time_value)

# 모델별 사용시간을 분 단위로 통일

new_use_time_list = []
for time in use_time_list:
    value = convert_time_minute(time)
    new_use_time_list.append(value)
    
print(len(new_use_time_list))
new_use_time_list[:10]

#### 무선청소기 흡입력 단위 통일시키기

suction_list

def get_suction(value):
    try:
        value = value.upper()
        if 'W' in value or 'AW' in value:
            result = value.replace('AW','').replace('W','')
            result = int(result.replace(',',''))
        elif 'PA' in value:
            result = value.replace('PA','')
            result = int(result.replace(',',''))/100
        else:
            result = None
            
        return result
    
    except:
        return None

# 함수 테스트

suction_list_test = ['100W', '10AW', '10,000Pa']
for power in suction_list_test:
    value = get_suction(power)
    print(power, value)

# 흡입력 단위 통일시키기
new_suction_list = []
for power in suction_list:
    value = get_suction(power)
    new_suction_list.append(value)
    
print(len(new_suction_list))
print(new_suction_list[:10])

#### 전처리결과 엑셀로 저장

pd_data = pd.DataFrame()
pd_data['카테고리'] = category_list
pd_data['회사명'] = company_list
pd_data['상품명'] = product_list
pd_data['가격'] = data['가격']
pd_data['사용시간'] = new_use_time_list
pd_data['흡입력'] = new_suction_list
pd_data.head()

pd_data['카테고리'].value_counts()

# 분석을 위해 핸디/스틱청소기만 추출
pd_data_final = pd_data[pd_data['카테고리'] == '핸디/스틱청소기']
print(pd_data_final.shape)
pd_data_final.head()

# 엑셀로 저장
pd_data_final.to_excel('crawling_data/6_2_danawa_data_final.xlsx', index=False)
